fix: Connect to device on device page
