+++ client/pages/index.vue
@@ -76,20 +76,8 @@ export default class PageIndex extends Vue {
     options?: any,
   ) => Promise<BottleView[]>
 
-  @Action('measurements/storeMeasurement') storeMeasurement: (
-    measurement: MeasurementCreate,
-  ) => Promise<MeasurementView>
-
-  mounted() {
-    this.fetchBottles().then(bottles =>
-      bottles.forEach(async bottle => {
-        const server = await connect(bottle.code)
-        watch(0x181d, 0x2a98, e => {
-          const weight = e.target.value.getUInt8(0) * 10
-          this.storeMeasurement({ bottleId: bottle.id, weight })
-        })
-      }),
-    )
+  async mounted() {
+    await this.fetchBottles()
 
     if (!('bluetooth' in navigator)) {
       this.toggleModal(true)
